improved hlsearch, see details
hlsearch can be annoying, especially after doing substitution command
where we don't really want to highlight the replaced pattern. therefore,
the hlsearch is now set to false by default. to compensate, S-m now
toggles the hlsearch on and off. statusline now show last searched term
if it is available in the current buffer